+++ src/app/onboarding/onboarding.module.ts
@@ -28,7 +28,7 @@ import { GlamArchiveTypeSelectComponent } from './components/glam/archive-type-s
 import { ArchiveTypeIconComponent } from './components/glam/archive-type-icon/archive-type-icon.component';
 import { CreateArchiveForMeScreenComponent } from './components/glam/create-archive-for-me-screen/create-archive-for-me-screen.component';
 import { FinalizeArchiveCreationScreenComponent } from './components/glam/finalize-archive-creation-screen/finalize-archive-creation-screen.component';
-import { GlamReasonsScreenComponent } from './components/glam-reasons-screen/glam-reasons-screen.component';
+import { GlamReasonsScreenComponent } from './components/glam/glam-reasons-screen/glam-reasons-screen.component';
 import { GlamGoalsScreenComponent } from './components/glam/glam-goals-screen/glam-goals-screen.component';
 import { GlamUserSurveySquareComponent } from './components/glam/glam-user-survey-square/glam-user-survey-square.component';
 import { OnboardingHeaderComponent } from './components/header/header.component';
